api_generator:cpp: support property bag via accessors - setProperty

+++ res/generators/templates/api/helpers/api_helper.rb
@@ -23,7 +23,7 @@ def api_generator_cpp_makeNativeType(gen_type)
     elsif gen_type == Rhogen::ApiGenerator::MethodParam::TYPE_HASH
         res = "rho::Hashtable<rho::StringW, rho::StringW>"
     elsif gen_type == Rhogen::ApiGenerator::MethodParam::TYPE_INT
-        res = "int"
+        res = "int64"
     elsif gen_type == Rhogen::ApiGenerator::MethodParam::TYPE_BOOL
         res = "bool"
     elsif gen_type == Rhogen::ApiGenerator::MethodParam::TYPE_DOUBLE
